Hardening (#176)
* Add network hardening

* more hardening

* update benches

* fmt

* clippy

* fixes

* clean

* fmt

* clean tests

* oredering

* fmt

* verify own share in PRE

* hash to sign scalar decaf377

* jwt skew tightened

* fix clippy

* revert decaf changes

* handle ring index better

* cleans

* fix nonce exhaustion

* timeout phases

* clean

+++ crates/crypto/benches/bls12_381/dkg.rs
@@ -8,7 +8,12 @@ pub struct Bls12381DkgBench;
 impl DkgBenchSetup for Bls12381DkgBench {
     type Node = DKGNode;
 
-    fn create_node(id: u32, threshold: usize, total_nodes: usize, session_id: u64) -> Box<DKGNode> {
+    fn create_node(
+        id: u32,
+        threshold: usize,
+        total_nodes: usize,
+        session_id: u128,
+    ) -> Box<DKGNode> {
         <DKGNode as Dkg>::new(id, threshold, total_nodes, session_id, DkgRole::Standard).unwrap()
     }
 }
