+++ crates/crypto/benches/decaf377/dkg.rs
@@ -8,7 +8,12 @@ pub struct Decaf377DkgBench;
 impl DkgBenchSetup for Decaf377DkgBench {
     type Node = DKGNode;
 
-    fn create_node(id: u32, threshold: usize, total_nodes: usize, session_id: u64) -> Box<DKGNode> {
+    fn create_node(
+        id: u32,
+        threshold: usize,
+        total_nodes: usize,
+        session_id: u128,
+    ) -> Box<DKGNode> {
         <DKGNode as Dkg>::new(id, threshold, total_nodes, session_id, DkgRole::Standard).unwrap()
     }
 }
